Multiple changes
It's been almost a year since I started working on this game engine. I started on November 19th 2022, about 8 months before I uploaded this project to GitHub, and I thought before this anniversary comes, I would add new features and fix some bugs. Here's what I've done so far:

- I made the physics a little bit better.
- I made the game rendering a bit more efficient. (Average 1-2 FPS increase from previous version)
- I added an experimental function called drawMemory, which draws the game's RAM in pixels onto the screen.

That's about it. :)

+++ SRC/TEXT.C
@@ -89,7 +89,7 @@ void displayAllText() // Prints everything text-related on screen
         {
             typeXPosition_Screen = 0;
             for (i = 0; i < typeXPosition_Text; i++)  // This focuses on the chat text cursor
-                typeXPosition_Screen += fontnumber_vwf[FONT_CHAT].charcode[type_line[i]];
+                typeXPosition_Screen += fontnumber_vwf[FONT_CHAT].charcode[(int)type_line[i]];
             drawimage(imgnumber[7], typeXPosition_Screen, 200);
         }
     }
